refactor : 패키지 분리 LottoCounter와 LottoGenerator를 컨트롤러에서 모델 패키지로 이동했습니다.
