+++ src/main/java/lotto/Model/LottoGenerator.java
@@ -1,4 +1,4 @@
-package lotto.Controller;
+package lotto.Model;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
