+++ src/test/java/lotto/LottoCounterTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.ArrayList;
 import java.util.List;
-import lotto.Controller.LottoCounter;
+import lotto.Model.LottoCounter;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
